+++ game/game_session.h
@@ -65,8 +65,8 @@ namespace msgame
             void SendMsg(std::vector<int32>& ids, msmessage::MessageBase *msg);
             void HandleMsg(int32 senderId, msmessage::MessageBase* msg);
 
-            void CalcMovePos(int64 senderId, btVector3 direction, int64 duration, btVector3& finalPos);
-            void AddJumpTask(int64 playerId, int64 jumpStart);
+            void CalcMovePos(int32 senderId, btVector3 direction, int64 duration, btVector3& finalPos);
+            void AddJumpTask(int32 playerId, int64 jumpStart);
 
             GameSession(int32 sessionId_);
             virtual ~GameSession();
